[iOS] [12673] Event Detail (Rate) - Allow user to rate event without leaving text feedback
Upon pressing the Rate button, the user is taken to the Rate Screen.
Currently, if the user selects a star rating and then completes the
rating by pressing "Send" button, the user will receive an error.
Allow the user to select the star rating and complete the rating
without having to leave a text review.