[iOS] [12800] Request App Store review (#108)
We should request app store reviews in the app.

We will prompt the user to review the app in the App Store after certain events occur (e.g. event rated, video finished playing, about screen loaded) and if the following requirements are met:

Running iOS 10.3 or above
App has been running for at least 5 minutes
It's been a week since we last prompted the user for a review.